add availability score to http module
