+++ www/http/blockchain.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/gorilla/mux"
 	"github.com/pactus-project/pactus/crypto/hash"
+	"github.com/pactus-project/pactus/types/vote"
 	pactus "github.com/pactus-project/pactus/www/grpc/gen/go"
 )
 
@@ -169,7 +170,37 @@ func (s *Server) writeValidatorTable(val *pactus.ValidatorInfo) *tableMaker {
 	tm.addRowInt("LastBondingHeight", int(val.LastBondingHeight))
 	tm.addRowInt("LastSortitionHeight", int(val.LastSortitionHeight))
 	tm.addRowInt("UnbondingHeight", int(val.UnbondingHeight))
+	tm.addRowInt("AvailabilityScore", int(val.AvailabilityScore))
 	tm.addRowBytes("Hash", val.Hash)
 
 	return tm
+}
+
+func (s *Server) ConsensusHandler(w http.ResponseWriter, _ *http.Request) {
+	res, err := s.blockchain.GetConsensusInfo(s.ctx,
+		&pactus.GetConsensusInfoRequest{})
+	if err != nil {
+		s.writeError(w, err)
+		return
+	}
+
+	tm := newTableMaker()
+	for i, cons := range res.Instances {
+		tm.addRowInt("== Validator", i+1)
+		tm.addRowValAddress("Address", cons.Address)
+		tm.addRowBool("Active", cons.Active)
+		tm.addRowInt("Height", int(cons.Height))
+		tm.addRowInt("Round", int(cons.Round))
+		tm.addRowString("Votes", "---")
+		for i, v := range cons.Votes {
+			tm.addRowInt("-- Vote #", i+1)
+			tm.addRowString("Type", vote.Type(v.Type).String())
+			tm.addRowString("Voter", v.Voter)
+			tm.addRowInt("Round", int(v.Round))
+			tm.addRowInt("CPRound", int(v.CpRound))
+			tm.addRowInt("CPValue", int(v.CpValue))
+			tm.addRowBlockHash("BlockHash", v.BlockHash)
+		}
+	}
+	s.writeHTML(w, tm.html())
 }
